+++ src/main/java/um/tesoreria/report/domain/dto/core/PlanDto.java
@@ -18,7 +18,7 @@ public class PlanDto {
     private Integer planId;
     private String nombre;
     
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssZ", timezone = "UTC")
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXX", timezone = "UTC")
     private OffsetDateTime fecha;
     
     private Byte publicar;